requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asyncio==4.0.0
attrs==26.1.0
audioop-lts==0.2.2
certifi==2026.2.25
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
jiter==0.13.0
magic-filter==1.0.12
multidict==6.7.1
numpy==2.4.4
openai==2.30.0
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python-dotenv==1.2.2
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
